requirements.txt
flask
python-nmap
python-owasp-zap-v2.4
requests
